Refactor: LetterBox 객체 분리
